+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.17] - 2021-04-26
+
+### Added
+
++ type Vectors with generic data storage
+
 ## [0.3.16] - 2021-04-26
 
 ### Added